Fix: お気に入りしていないレシピはお気に入り解除できなくした #60

+++ app/controllers/favorites_controller.rb
@@ -2,7 +2,7 @@ class FavoritesController < ApplicationController
   def create
     # gurestなら
     if current_user.guest?
-      redirect_to result_path, notice: 'ログインしてください'
+      redirect_to result_path, alert: 'ログインしてください'
     else 
       # guest以外なら
       favorite = current_user.favorites.build(recipe_id: params[:recipe_id])
@@ -14,8 +14,13 @@ class FavoritesController < ApplicationController
 
   def destroy
     # binding.pry
-    favorite = current_user.favorites.find_by(recipe_id: params[:id])
-    favorite.destroy!
-    redirect_to result_path, notice: 'お気に入り解除しました'
+    # とりあえずお気に入り登録していないレシピはエラーにならないようにした
+    if current_user.favorites.find_by(recipe_id: params[:id])
+      favorite = current_user.favorites.find_by(recipe_id: params[:id])
+      favorite.destroy!
+      redirect_to result_path, notice: 'お気に入り解除しました'
+    else
+      redirect_to result_path, alert: 'お気に入りしていません'
+    end
   end
 end
